[fix] EventReview 분리 및 라이브러리 추가 공연/정보 상세 페이지를 총괄하는 Detail.vue를 Event.vue로 이름 변경 및 한줄평 남기는 Review를 따로 분리
