+++ package.json
@@ -11,7 +11,9 @@
   "dependencies": {
     "@tailwindcss/line-clamp": "^0.4.4",
     "axios": "^1.8.4",
+    "dompurify": "^3.2.5",
     "jwt-decode": "^4.0.0",
+    "marked": "^15.0.8",
     "pinia": "^3.0.1",
     "pinia-plugin-persistedstate": "^4.2.0",
     "v3-infinite-loading": "^1.3.2",
