+++ src/assets/data/review.js
@@ -1,0 +1,38 @@
+export default [
+    {
+        idx: 1,
+        writer: '별자리',
+        date: '2025.04.18',
+        content: '햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!'
+    },
+    {
+        idx: 2,
+        writer: '별이',
+        date: '2025.04.18',
+        content: '배경 음악이 정말 몰입감을 높여줬어요. 다음에도 꼭 볼 거예요!'
+    },
+    {
+        idx: 3,
+        writer: '별나라',
+        date: '2025.04.18',
+        content: '무대 조명과 의상이 환상적이었어요.'
+    },
+    {
+        idx: 4,
+        writer: '별주부전',
+        date: '2025.04.18',
+        content: '배우들의 연기력이 너무 좋았어요!'
+    },
+    {
+        idx: 5,
+        writer: '별도달도따다줄게',
+        date: '2025.04.18',
+        content: '대사 하나하나가 가슴을 울렸어요.'
+    },
+    {
+        idx: 6,
+        writer: '별별별',
+        date: '2025.04.18',
+        content: '스토리가 탄탄해서 끝까지 몰입했어요.'
+    }
+]
