+++ src/events/EventReview.vue
@@ -1,246 +1,96 @@
 <script setup>
+import { ref, computed, onMounted } from 'vue'
+import reviews from '@/assets/data/review'
 
+const currentPage = ref(1)
+const itemsPerPage = 10
+
+// 페이지네이션 계산
+const totalPages = computed(() =>
+  Math.ceil(reviews.length / itemsPerPage)
+)
+
+const paginatedReviews = computed(() =>
+  reviews.slice((currentPage.value - 1) * itemsPerPage, currentPage.value * itemsPerPage)
+)
+
+const goToPage = (page) => {
+  if (page >= 1 && page <= totalPages.value) {
+    currentPage.value = page
+  }
+}
 </script>
 
 <template>
-  <div data-layer="Frame 1321317502" class="inline-flex flex-col justify-start items-start gap-8">
-    <div data-layer="REVIEW LIST" class="self-stretch flex flex-col justify-start items-start gap-4">
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
+  <section class="flex flex-col gap-8">
+    <h3 class="text-xl font-bold text-neutral-800">한줄평 ({{ reviews.length }})</h3>
+
+    <!-- 리뷰 카드 리스트 -->
+    <div class="flex flex-col gap-4">
+      <div
+        v-for="review in paginatedReviews"
+        :key="review.idx"
+        class="self-stretch px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.5px] outline-violet-50 flex flex-col justify-start items-start gap-2.5"
+      >
+        <div class="inline-flex justify-start items-center gap-4">
+          <div class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex justify-center items-center">
+            <div class="text-center text-violet-600 text-xs font-bold">별</div>
           </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별자리" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별자리</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
+          <div class="w-full inline-flex flex-col gap-1">
+            <div class="flex justify-between items-center">
+              <div class="flex items-end gap-3">
+                <div class="text-neutral-800 text-base font-bold">{{ review.writer }}</div>
+                <div class="text-yellow-400 text-base">★★★★★</div>
               </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
+              <div class="text-neutral-400 text-xs">{{ review.date }}</div>
             </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
-          </div>
-        </div>
-      </div>
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
-          </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별이" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별이</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
-              </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
-            </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
-          </div>
-        </div>
-      </div>
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
-          </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별나라" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별나라</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
-              </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
-            </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
-          </div>
-        </div>
-      </div>
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
-          </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별주부전" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별주부전</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
-              </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
-            </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
-          </div>
-        </div>
-      </div>
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
-          </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별도달도따다줄게" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별도달도따다줄게</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
-              </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
-            </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
-          </div>
-        </div>
-      </div>
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
-          </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별별별" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별별별</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
-              </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
-            </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
-          </div>
-        </div>
-      </div>
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
-          </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별따먹자" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별따먹자</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
-              </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
-            </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
-          </div>
-        </div>
-      </div>
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
-          </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별주부와토끼" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별주부와토끼</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
-              </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
-            </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
-          </div>
-        </div>
-      </div>
-      <div data-layer="REVIEW ITEM" class="self-stretch h-20 px-6 py-5 bg-white rounded-[10px] outline outline-1 outline-offset-[-0.50px] outline-violet-50 flex flex-col justify-start items-start gap-2.5">
-        <div data-layer="Frame 1321317485" class="inline-flex justify-start items-center gap-4">
-          <div data-layer="Frame 1321317480" class="w-9 h-10 px-3.5 py-3 bg-violet-50 rounded-[20px] inline-flex flex-col justify-center items-center gap-2.5">
-            <div data-layer="별" class="text-center justify-start text-violet-600 text-xs font-bold font-['Inter']">별</div>
-          </div>
-          <div data-layer="Frame 1321317484" class="w-[1094.45px] inline-flex flex-col justify-start items-start gap-1">
-            <div data-layer="Frame 1321317483" class="self-stretch inline-flex justify-between items-center">
-              <div data-layer="Frame 1321317482" class="flex justify-start items-end gap-3">
-                <div data-layer="별이반짝" class="justify-start text-neutral-800 text-base font-bold font-['Inter']">별이반짝</div>
-                <div data-layer="★★★★★" class="justify-start text-yellow-400 text-base font-normal font-['Inter']">★★★★★</div>
-              </div>
-              <div data-layer="2025.04.18" class="justify-start text-neutral-400 text-xs font-normal font-['Inter']">2025.04.18</div>
-            </div>
-            <div data-layer="햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!" class="self-stretch justify-start text-neutral-800 text-sm font-normal font-['Inter']">햄릿의 독백 장면이 압권이었습니다. 배우의 감정 표현이 정말 섬세했어요!</div>
+            <div class="text-neutral-800 text-sm">{{ review.content }}</div>
           </div>
         </div>
       </div>
     </div>
-    <div data-layer="Frame 1321317503" class="w-[1200px] h-10 relative">
-      <div data-layer="페이지 이동 버튼 그룹" class="left-[367px] top-0 absolute inline-flex justify-start items-center gap-2.5">
-        <div data-svg-wrapper data-layer="page before btn" class="relative">
-          <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            <path d="M26 10.936L16 20.936L26 30.936" stroke="#495057" stroke-width="2"/>
+
+    <!-- 페이지네이션 -->
+    <section class="flex justify-center">
+      <div class="h-10 flex justify-center items-center gap-2.5">
+        <!-- 이전 버튼 -->
+        <button @click="goToPage(currentPage - 1)" :disabled="currentPage === 1">
+          <svg width="42" height="42" viewBox="0 0 42 42" fill="none">
+            <path d="M37 1H5C2.8 1 1 2.8 1 5v32c0 2.2 1.8 4 4 4h32c2.2 0 4-1.8 4-4V5c0-2.2-1.8-4-4-4Z" fill="white" stroke="#DEE2E6"/>
+            <path d="M26 11L16 21l10 10" stroke="#495057" stroke-width="2"/>
           </svg>
-        </div>
-        <div data-layer="page1 btn" class="w-10 h-10 relative">
-          <div data-svg-wrapper data-layer="Vector" class="left-0 top-0 absolute">
-            <svg width="40" height="41" viewBox="0 0 40 41" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M36 0.936035H4C1.79086 0.936035 0 2.7269 0 4.93604V36.936C0 39.1452 1.79086 40.936 4 40.936H36C38.2091 40.936 40 39.1452 40 36.936V4.93604C40 2.7269 38.2091 0.936035 36 0.936035Z" fill="#7048E8"/>
-            </svg>
-          </div>
-          <div data-layer="1" class="left-[16px] top-[9px] absolute text-center justify-start text-white text-base font-bold font-['Inter']">1</div>
-        </div>
-        <div data-layer="page2 btn" class="w-10 h-10 relative">
-          <div data-svg-wrapper data-layer="Vector" class="left-0 top-0 absolute">
-            <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            </svg>
-          </div>
-          <div data-layer="2" class="left-[15px] top-[10px] absolute text-center justify-start text-zinc-600 text-base font-normal font-['Inter']">2</div>
-        </div>
-        <div data-layer="page3 btn" class="w-10 h-10 relative">
-          <div data-svg-wrapper data-layer="Vector" class="left-0 top-0 absolute">
-            <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            </svg>
-          </div>
-          <div data-layer="3" class="left-[15px] top-[10px] absolute text-center justify-start text-zinc-600 text-base font-normal font-['Inter']">3</div>
-        </div>
-        <div data-layer="page4 btn" class="w-10 h-10 relative">
-          <div data-svg-wrapper data-layer="Vector" class="left-0 top-0 absolute">
-            <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            </svg>
-          </div>
-          <div data-layer="4" class="left-[15px] top-[10px] absolute text-center justify-start text-zinc-600 text-base font-normal font-['Inter']">4</div>
-        </div>
-        <div data-layer="page5 btn" class="w-10 h-10 relative">
-          <div data-svg-wrapper data-layer="Vector" class="left-0 top-0 absolute">
-            <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            </svg>
-          </div>
-          <div data-layer="5" class="left-[15px] top-[10px] absolute text-center justify-start text-zinc-600 text-base font-normal font-['Inter']">5</div>
-        </div>
-        <div data-layer="page6 btn" class="w-10 h-10 relative">
-          <div data-svg-wrapper data-layer="Vector" class="left-0 top-0 absolute">
-            <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            </svg>
-          </div>
-          <div data-layer="6" class="left-[15px] top-[10px] absolute text-center justify-start text-zinc-600 text-base font-normal font-['Inter']">6</div>
-        </div>
-        <div data-layer="page7 btn" class="w-10 h-10 relative">
-          <div data-svg-wrapper data-layer="Vector" class="left-0 top-0 absolute">
-            <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            </svg>
-          </div>
-          <div data-layer="7" class="left-[15px] top-[10px] absolute text-center justify-start text-zinc-600 text-base font-normal font-['Inter']">7</div>
-        </div>
-        <div data-svg-wrapper data-layer="page after btn" class="relative">
-          <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            <path d="M16 10.936L26 20.936L16 30.936" stroke="#495057" stroke-width="2"/>
+        </button>
+
+        <!-- 번호 버튼 -->
+        <button
+          v-for="page in totalPages"
+          :key="page"
+          @click="goToPage(page)"
+          class="relative w-10 h-10"
+        >
+          <svg :width="42" :height="42" viewBox="0 0 42 42" fill="none">
+            <path
+              d="M37 1H5C2.8 1 1 2.8 1 5v32c0 2.2 1.8 4 4 4h32c2.2 0 4-1.8 4-4V5c0-2.2-1.8-4-4-4Z"
+              :fill="page === currentPage ? '#7048E8' : 'white'"
+              :stroke="page === currentPage ? '#7048E8' : '#DEE2E6'"
+            />
           </svg>
-        </div>
-        <div data-svg-wrapper data-layer="page far after btn" class="relative">
-          <svg width="42" height="42" viewBox="0 0 42 42" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M37 0.936035H5C2.79086 0.936035 1 2.7269 1 4.93604V36.936C1 39.1452 2.79086 40.936 5 40.936H37C39.2091 40.936 41 39.1452 41 36.936V4.93604C41 2.7269 39.2091 0.936035 37 0.936035Z" fill="white" stroke="#DEE2E6"/>
-            <path d="M21 10.936L31 20.936L21 30.936" stroke="#495057" stroke-width="2"/>
-            <path d="M11 10.936L21 20.936L11 30.936" stroke="#495057" stroke-width="2"/>
+          <div
+            class="absolute left-[15px] top-[10px] text-base text-center"
+            :class="page === currentPage ? 'text-white font-bold' : 'text-zinc-600 font-normal'"
+          >
+            {{ page }}
+          </div>
+        </button>
+
+        <!-- 다음 버튼 -->
+        <button @click="goToPage(currentPage + 1)" :disabled="currentPage === totalPages">
+          <svg width="42" height="42" viewBox="0 0 42 42" fill="none">
+            <path d="M37 1H5C2.8 1 1 2.8 1 5v32c0 2.2 1.8 4 4 4h32c2.2 0 4-1.8 4-4V5c0-2.2-1.8-4-4-4Z" fill="white" stroke="#DEE2E6"/>
+            <path d="M16 11l10 10-10 10" stroke="#495057" stroke-width="2"/>
           </svg>
-        </div>
+        </button>
       </div>
-    </div>
-  </div>
+    </section>
+  </section>
 </template>
-
-<style scoped>
-
-</style>
